requirements.txt
streamlit
google-generativeai
rapidfuzz
python-dotenv